Refactor hop addition action handler
The `HopAddition.Delete` action is unique in the sense that it doesn't
change `HopAddition.Model`, so we don't call `HopAddition.update` from
the parent's `update` function. It is handled by the parent, which
removes the hop addition from its list.

This commit adds a `DeleteHopAddition` action in the parent component,
which doesn't take a `HopAddition.Msg` as an argument, since it doesn't
need to call `HopAddition.update`.

By mapping to two different parent messages, we can get rid of the
nested case statement in the parent's update function.