Use boolean to track hop addition deletion
This commit adds a boolean property to the hop addition model to track
if it's deleted. This way, the parent component doesn't have to know
about the `HopAddition.Delete` message. Instead, the parent update
function calls the `HopAddition.update` function, which may in turn set
`isDeleted` on the `HopAddition` model. Then, the parent update function
filters out hop additions where `isDeleted` is `True`.